Use a separate cache when proxyquireify is active
Instead of entirely disabling the browserify cache, just switch to
a different cache each time `proxyquire` is called. This allows
modules like `tape`, which have global state and/or circular
dependencies, to work correctly.

Also, move the require monkeypatching from the transform to the
prelude, wich makes it possible to build code with proxyquireify
even if proxyquireify is not used and is not built into the
bundle. This allows test code to be always built with proxyquireify
even if the module is not included as a dependency.